backend/requirements.txt
setuptools>=65.0.0
wheel>=0.38.0
pip>=23.0.0
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.0.5
numpy==1.19.3
pandas==1.1.4
python-dotenv==0.15.0
requests==2.31.0
scikit-learn==0.24.2
scipy==1.5.4
tqdm==4.54.1
urllib3==1.26.4
PyJWT==2.8.0
pymongo==4.6.0
dnspython==2.4.2
gunicorn==21.2.0
certifi==2023.7.22
# Compatibility fixes
# Flask 2.3.3 manages these dependencies automatically:
# - Jinja2 (requires >=3.1.2)
# - MarkupSafe
# - itsdangerous
# Let Flask handle these to avoid conflicts
Werkzeug==2.3.7
# SQLAlchemy version managed by Flask-SQLAlchemy 3.0.5
